fix(admin): guarda acceso a /admin solo para rol Administrador (server guard leyendo cookie tm_auth)
